% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmasgifts.Rmd.
\name{make_phrase}
\alias{make_phrase}
\title{Create the gift phrase for a specific day}
\usage{
make_phrase(day_num, item, verb, adjective, location)
}
\arguments{
\item{day_num}{Number of the day in digits}

\item{item}{Name of the gift item}

\item{verb}{Name of the action the gift item}

\item{adjective}{Adjective to describe the gift item}

\item{location}{Location of the gift item}
}
\description{
The first step in writing the 12 Days of Christmas song is to pluralize the gift items.
}