% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmasgifts.Rmd.
\name{pluralize_gift}
\alias{pluralize_gift}
\title{Pluralize a gift item}
\usage{
pluralize_gift(item)
}
\arguments{
\item{item}{Name of a gift item}
}
\description{
The first step in writing the 12 Days of Christmas song is to pluralize the gift items.
}